Changes for comments from @gwenshap and @theduderog
Remove the Connectivity checking code as it was not doing anything. I
think I did that because the python code had that. As @gwenshap
mentioned the ZK Client code retires when the server is unavailable.
The Kafka client fails but my code retries anyways, so that is also
fine. The other condition that I though it guarded against was when a
host was not resolvable but that  will never happen in properly
configured docker/kubernetes environments (the DNS will always resolve
even if the containers are not ready) and the check should fail if
DNS/hostnames is not configured properly. And all the tests pass so I
don’t think it is required.