Make changes as per @cotedm comments.
Fix docs
Fix logging
Removed the properties as I have a command to generate it now :)
Remove the zktimeout multiplier. I realized that I did that to debug a
issue and forgot to remove it. If the ZK session times ou t during an
operation, I think it is ok to fail as zkTimeout = timeout for the
entire command set by the user.