Fixed flaky tests and make ZK SASL tests work.
- The SSL tests for Kafka are flaky because the consumer rebalance take
longer than docker-compose timeout sometimes. Replaced console consumer
with Kafkacat.

- ZK SASL : The zkClient somehow gets all IPs associated with the
hostnames for a given connect string .

In this, case localhost was resolved to the VM hostname (as docker host
networking just uses the host n/w stack). This causes problems with
SASL as principal names are tied to hostnames. So, I replaced localhost
with IP (127.0.0.1) and also changed localhost in the principal names
to IP (127.0.0.1)

Note that this lookup also causes issues with the bridged n/w and I
could not find a workaround for it.