Support SASL for Zookeeper. (#79)
* Config tests pass.

* Replace cub with Java version + SASL tests.

* Kafka + ZK SASL tests pass.

* Build docker-utils.

* Add inline docs for cub and dub.

Fixes #42.

* Add java code from docker-utils.

* Support zk for kafka-ready.

* ZK, Kafka, SR tests run.

* Rest Proxy and connect tests work.

* Control center tests pass !

* Cleanup and tests for cub.

* Some more inline docs.

* Ignore hacking code.

* Changes for comments from @gwenshap and @theduderog

Remove the Connectivity checking code as it was not doing anything. I
think I did that because the python code had that. As @gwenshap
mentioned the ZK Client code retires when the server is unavailable.
The Kafka client fails but my code retries anyways, so that is also
fine. The other condition that I though it guarded against was when a
host was not resolvable but that  will never happen in properly
configured docker/kubernetes environments (the DNS will always resolve
even if the containers are not ready) and the check should fail if
DNS/hostnames is not configured properly. And all the tests pass so I
don’t think it is required.

* Make changes as per @cotedm comments.

Fix docs
Fix logging
Removed the properties as I have a command to generate it now :)
Remove the zktimeout multiplier. I realized that I did that to debug a
issue and forgot to remove it. If the ZK session times ou t during an
operation, I think it is ok to fail as zkTimeout = timeout for the
entire command set by the user.

* Fixed typo.

* Fixes for @cotedm comments.

* Integrate tests into the makefile.

* Refactor test-all make target.

* Fix bug (KAFKA_OPTS will never be passed).

* Fix grammatical errors in logging messages.

* Fixed flaky tests and make ZK SASL tests work.

- The SSL tests for Kafka are flaky because the consumer rebalance take
longer than docker-compose timeout sometimes. Replaced console consumer
with Kafkacat.

- ZK SASL : The zkClient somehow gets all IPs associated with the
hostnames for a given connect string .

In this, case localhost was resolved to the VM hostname (as docker host
networking just uses the host n/w stack). This causes problems with
SASL as principal names are tied to hostnames. So, I replaced localhost
with IP (127.0.0.1) and also changed localhost in the principal names
to IP (127.0.0.1)

Note that this lookup also causes issues with the bridged n/w and I
could not find a workaround for it.

* Fixes issue with SASL + ZKClient.

The ZKClient needs a hostname that supports both DNS->IP and
IP->Lookups. To make sure this condition is met,  I created a
/etc/hosts that maps to eth0.